added kbuild, moved code to src/, and make to build/ directory
