+++ Kbuild
@@ -1,0 +1,2 @@
+obj-m += tmem_csc450.o
+tmem_csc450-y += src/main.o
